requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.3.3
entrypoints==0.3
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
yarl==1.4.2
